requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0

llama-cpp-python>=0.2.0

faster-whisper>=1.0.0
openai-whisper>=20231117

sounddevice>=0.4.6
numpy>=1.24.0
webrtcvad>=2.0.10 

huggingface-hub>=0.20.0

psutil>=5.9.0

qdarktheme>=2.1.0

colorlog>=6.7.0

pandas>=2.0.0
openpyxl>=3.1.0

scipy>=1.11.0

pyinstaller>=5.13.0
auto-py-to-exe>=2.40.0